Week8_Lab/requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
trl>=0.8.0
groq>=0.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
scikit-learn>=1.4.0
numpy>=1.24.0
reportlab>=4.0
